chore: Add menuScan.dbd file, update .gitignore, and add test.db
This commit adds the menuScan.dbd file to the /data/epics/dbd directory and updates the .gitignore file to include the new file. It also adds the test.db file to the iocs/scanTest/db directory. These changes enhance the functionality and configuration of the EPICS Ethercat Base Docker image.